# openapi.schema.json
{"openapi":"3.0.0","info":{"title":"words","version":"0.0.1","description":""},"paths":{"/test":{"get":{"tags":["App"],"responses":{"200":{"content":{"text/plain":{"schema":{"type":"string"}}},"description":"ok"}},"parameters":[]}},"":{"get":{"tags":["App"],"responses":{"200":{"content":{"text/plain":{"schema":{"type":"string"}}},"description":"ok"}},"parameters":[]}},"/test1":{"post":{"tags":["App"],"responses":{"200":{"content":{"text/plain":{"schema":{"type":"string"}}},"description":"ok"}},"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/App"}}}},"parameters":[]}},"/auth/login":{"post":{"tags":["Auth"],"responses":{"200":{"content":{"application/json":{}},"description":"ok"}},"parameters":[]}},"/cats":{"post":{"tags":["Cats"],"responses":{"200":{"content":{"application/json":{}},"description":"ok"}},"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cat"}}}},"parameters":[]},"get":{"tags":["Cats"],"responses":{"200":{"content":{"application/json":{"schema":{"type":"array","items":{}}}},"description":"ok"}},"parameters":[]}},"/cats/{id}":{"get":{"tags":["Cats"],"responses":{"200":{"content":{"application/json":{"schema":{"type":"array","items":{}}}},"description":"ok"}},"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}]},"delete":{"tags":["Cats"],"responses":{"200":{"content":{"application/json":{"schema":{"type":"array","items":{}}}},"description":"ok"}},"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}]}},"/users/create":{"post":{"tags":["User"],"responses":{"200":{"content":{"application/json":{}},"description":"ok"}},"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"parameters":[]}}},"components":{"schemas":{"App":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"imageUrl":{"type":"string"},"category":{"type":"string"},"meniTester":{"type":"object","properties":{}},"validate":{"type":"object","properties":{}},"toJSON":{"type":"object","properties":{}},"clone":{"type":"object","properties":{}}},"required":["id","name","description","price","imageUrl","category","meniTester","validate","toJSON","clone"]},"Cat":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"age":{"type":"number"},"breed":{"type":"string"},"validate":{"type":"object","properties":{}},"toJSON":{"type":"object","properties":{}},"clone":{"type":"object","properties":{}}},"required":["name","age","breed","validate","toJSON","clone"]},"User":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"},"validate":{"type":"object","properties":{}},"toJSON":{"type":"object","properties":{}},"clone":{"type":"object","properties":{}}},"required":["id","firstName","lastName","email","phone","password","validate","toJSON","clone"]}}},"tags":[{"name":"App"},{"name":"Auth"},{"name":"Cats"},{"name":"User"}]}